Extract shared catalog module from TUI frontend
Move RomEntry and catalog logic from src/frontends/tui/ to
src/platform/catalog/ so both TUI and native graphical frontends
can share the ROM discovery and enrichment code.

Extend RomEntry with metadata fields (metadata_game_id, genres,
overview, release_date, players, rating, boxart_path,
screenshot_paths, is_favorite) in preparation for the ROM
navigation UI (#2335).

TUI modules now re-export from the shared catalog module.
All 39 TUI tests and 19 shared catalog tests pass.
