+++ src/frontends/tui/app.rs
@@ -329,6 +329,15 @@ mod tests {
                 hardware: Some("NES NTSC".to_string()),
                 crc: Some("DEADBEEF".to_string()),
                 recording_duration: None,
+                metadata_game_id: None,
+                genres: Vec::new(),
+                overview: None,
+                release_date: None,
+                players: None,
+                rating: None,
+                boxart_path: None,
+                screenshot_paths: Vec::new(),
+                is_favorite: false,
             })
             .collect()
     }
